+++ veterans/schema.ts
@@ -1,10 +1,12 @@
 import { User } from './keystone/models/User';
 import { Initiative } from './keystone/models/Initiative';
 import { Category, Source } from './keystone/models/CommonTaxonomyModel';
+import { Region } from './keystone/models/Region';
 
 export const lists = {
   User,
   Initiative,
   Category,
   Source,
+  Region,
 };